Moved type info tests back into a seperate file.
