--- internal/expr/export_test.go
@@ -1,3 +1,0 @@
-package expr
-
-var TypeInfo = typeInfo
