+++ internal/expr/expr_test.go
@@ -1,7 +1,6 @@
 package expr_test
 
 import (
-	"sync"
 	"testing"
 
 	"github.com/canonical/sqlair/internal/expr"
@@ -297,133 +296,4 @@ func FuzzParser(f *testing.F) {
 		parser := expr.NewParser()
 		parser.Parse(s)
 	})
-}
-
-func (s *ExprSuite) TestReflectSimpleConcurrent(c *C) {
-	type myStruct struct{}
-
-	// Get the type info of a struct sequentially.
-	var seqSt myStruct
-	seqInfo, err := expr.TypeInfo(seqSt)
-	c.Assert(err, IsNil)
-
-	// Get some type info concurrently.
-	var concSt myStruct
-	wg := sync.WaitGroup{}
-
-	// Set up some concurrent access.
-	for i := 0; i < 5; i++ {
-		wg.Add(1)
-		go func() {
-			_, _ = expr.TypeInfo(concSt)
-			wg.Done()
-		}()
-	}
-
-	// Get type info alongside concurrent access.
-	concInfo, err := expr.TypeInfo(concSt)
-	c.Assert(err, IsNil)
-
-	c.Assert(seqInfo, Equals, concInfo)
-
-	wg.Wait()
-}
-
-func (s *ExprSuite) TestReflectNonStructType(c *C) {
-	type mymap map[int]int
-	var nonStructs = []any{
-		mymap{},
-		int(0),
-		string(""),
-		map[string]string{},
-	}
-
-	for _, value := range nonStructs {
-		i, err := expr.TypeInfo(value)
-		c.Assert(err, ErrorMatches, "internal error: attempted to obtain struct information for something that is not a struct: .*")
-		c.Assert(i, IsNil)
-	}
-}
-
-func (s *ExprSuite) TestReflectBadTagError(c *C) {
-
-	var unsupportedFlag = []any{
-		struct {
-			ID int64 `db:"id,bad-juju"`
-		}{99},
-		struct {
-			ID int64 `db:","`
-		}{99},
-		struct {
-			ID int64 `db:"id,omitempty,ddd"`
-		}{99},
-	}
-
-	var tagEmpty = []any{
-		struct {
-			ID int64 `db:",omitempty"`
-		}{99},
-	}
-
-	var invalidColumn = []any{
-		struct {
-			ID int64 `db:"5id"`
-		}{99},
-		struct {
-			ID int64 `db:"+id"`
-		}{99},
-		struct {
-			ID int64 `db:"-id"`
-		}{99},
-		struct {
-			ID int64 `db:"id/col"`
-		}{99},
-		struct {
-			ID int64 `db:"id$$"`
-		}{99},
-		struct {
-			ID int64 `db:"id|2005"`
-		}{99},
-		struct {
-			ID int64 `db:"id|2005"`
-		}{99},
-	}
-
-	for _, value := range unsupportedFlag {
-		_, err := expr.TypeInfo(value)
-		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: unsupported flag .*")
-	}
-	for _, value := range tagEmpty {
-		_, err := expr.TypeInfo(value)
-		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: .*")
-	}
-	for _, value := range invalidColumn {
-		_, err := expr.TypeInfo(value)
-		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: invalid column name in 'db' tag: .*")
-	}
-}
-
-func (s *ExprSuite) TestReflectValidTag(c *C) {
-	var validTags = []any{
-		struct {
-			ID int64 `db:"id_"`
-		}{99},
-		struct {
-			ID int64 `db:"id5"`
-		}{99},
-		struct {
-			ID int64 `db:"_i_d_55"`
-		}{99},
-		struct {
-			ID int64 `db:"id_2002"`
-		}{99},
-		struct {
-			ID int64 `db:"IdENT99"`
-		}{99},
-	}
-
-	for _, value := range validTags {
-		_, err := expr.TypeInfo(value)
-		c.Assert(err, IsNil)
-	}
 }
