+++ internal/expr/typeinfo_test.go
@@ -2,6 +2,7 @@ package expr
 
 import (
 	"reflect"
+	"sync"
 	"testing"
 
 	. "gopkg.in/check.v1"
@@ -44,4 +45,133 @@ func (e *ExprInternalSuite) TestReflectStruct(c *C) {
 	c.Assert(ok, Equals, true)
 	c.Assert("Name", Equals, name.name)
 	c.Assert(name.omitEmpty, Equals, true)
+}
+
+func (s *ExprInternalSuite) TestReflectSimpleConcurrent(c *C) {
+	type myStruct struct{}
+
+	// Get the type info of a struct sequentially.
+	var seqSt myStruct
+	seqInfo, err := typeInfo(seqSt)
+	c.Assert(err, IsNil)
+
+	// Get some type info concurrently.
+	var concSt myStruct
+	wg := sync.WaitGroup{}
+
+	// Set up some concurrent access.
+	for i := 0; i < 5; i++ {
+		wg.Add(1)
+		go func() {
+			_, _ = typeInfo(concSt)
+			wg.Done()
+		}()
+	}
+
+	// Get type info alongside concurrent access.
+	concInfo, err := typeInfo(concSt)
+	c.Assert(err, IsNil)
+
+	c.Assert(seqInfo, Equals, concInfo)
+
+	wg.Wait()
+}
+
+func (s *ExprInternalSuite) TestReflectNonStructType(c *C) {
+	type mymap map[int]int
+	var nonStructs = []any{
+		mymap{},
+		int(0),
+		string(""),
+		map[string]string{},
+	}
+
+	for _, value := range nonStructs {
+		i, err := typeInfo(value)
+		c.Assert(err, ErrorMatches, "internal error: attempted to obtain struct information for something that is not a struct: .*")
+		c.Assert(i, IsNil)
+	}
+}
+
+func (s *ExprInternalSuite) TestReflectBadTagError(c *C) {
+
+	var unsupportedFlag = []any{
+		struct {
+			ID int64 `db:"id,bad-juju"`
+		}{99},
+		struct {
+			ID int64 `db:","`
+		}{99},
+		struct {
+			ID int64 `db:"id,omitempty,ddd"`
+		}{99},
+	}
+
+	var tagEmpty = []any{
+		struct {
+			ID int64 `db:",omitempty"`
+		}{99},
+	}
+
+	var invalidColumn = []any{
+		struct {
+			ID int64 `db:"5id"`
+		}{99},
+		struct {
+			ID int64 `db:"+id"`
+		}{99},
+		struct {
+			ID int64 `db:"-id"`
+		}{99},
+		struct {
+			ID int64 `db:"id/col"`
+		}{99},
+		struct {
+			ID int64 `db:"id$$"`
+		}{99},
+		struct {
+			ID int64 `db:"id|2005"`
+		}{99},
+		struct {
+			ID int64 `db:"id|2005"`
+		}{99},
+	}
+
+	for _, value := range unsupportedFlag {
+		_, err := typeInfo(value)
+		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: unsupported flag .*")
+	}
+	for _, value := range tagEmpty {
+		_, err := typeInfo(value)
+		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: .*")
+	}
+	for _, value := range invalidColumn {
+		_, err := typeInfo(value)
+		c.Assert(err, ErrorMatches, "cannot parse tag for field .ID: invalid column name in 'db' tag: .*")
+	}
+}
+
+func (s *ExprInternalSuite) TestReflectValidTag(c *C) {
+	var validTags = []any{
+		struct {
+			ID int64 `db:"id_"`
+		}{99},
+		struct {
+			ID int64 `db:"id5"`
+		}{99},
+		struct {
+			ID int64 `db:"_i_d_55"`
+		}{99},
+		struct {
+			ID int64 `db:"id_2002"`
+		}{99},
+		struct {
+			ID int64 `db:"IdENT99"`
+		}{99},
+	}
+
+	for _, value := range validTags {
+		_, err := typeInfo(value)
+		c.Assert(err, IsNil)
+	}
 }
